Polish code for TestCase.
